add manage delte

+++ src/pages/ManageInventoryItems/ManageInventoryItems.js
@@ -13,38 +13,37 @@ import './ManageInventoryItems.css';
 const ManageInventoryItems = () => {
    
   const [products,setProducts]=useData();
-  
   const [phones, setPhones]=useProduct();
  const {_id}= phones;
 
   
   const handleDelete=id=>{
-   
-    fetch(`http://localhost:5000/inventory/${id}`,{
-      method:"DELETE",
-    })
-    .then(res=>res.json())
-    .then(data =>console.log(data));
-  }
+    const proceed=window.confirm("Are u sure?");
+    if(proceed){
+       console.log(id);
+       fetch(`http://localhost:5000/inventory/${id}`, {
+         method: "DELETE",
+       })
+         .then((res) => res.json())
+         .then((data) => console.log(data));
 
+    }
+
+    }
+   
+   
   const columns = [
     {
-      title: "Id",
-      dataIndex: "Id",
-      key: "_Id",
+      title: "productName",
+      dataIndex: "productName",
+      key: "productName",
       width: 100,
     },
     {
-      title: "ProductName",
-      dataIndex: "ProductName",
-      key: "ProductName",
-      width: 100,
-    },
-    {
-      title: "quantity",
-      dataIndex: "quantity",
-      key: "quantity",
-      width: 100,
+      title: "supplierName",
+      dataIndex: "supplierName",
+      key: "supplierName",
+      width: 200,
     },
     {
       title: "price",
@@ -53,10 +52,10 @@ const ManageInventoryItems = () => {
       width: 200,
     },
     {
-      title: "supplier",
-      dataIndex: "supplier",
-      key: "supplier",
-      width: 200,
+      title: "quantity",
+      dataIndex: "quantity",
+      key: "quantity",
+      width: 100,
     },
 
     {
@@ -64,7 +63,7 @@ const ManageInventoryItems = () => {
       dataIndex: "",
       key: "operations",
       render: () => (
-        <a href="#">
+        <a href="/manageinventory">
           <button onClick={() => handleDelete(_id)} className="trash-btn">
             <img src={trash} alt="" />
           </button>
@@ -85,6 +84,8 @@ const ManageInventoryItems = () => {
             Add new item
           </a>
         </Button>
+
+        
       </div>
     </div>
   );
